extract parsestring
